[feature/styles] - move styles into thei own files
