+++ src/components/input/Input_styles.js
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const StInput = styled.input`
+  outline: none;
+  padding: 10px 60px;
+`;
+
+export {
+  StInput
+}
