+++ src/components/pokemon/PokemonDetail_styles.js
@@ -1,24 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
 import styled from '@emotion/styled'
-
 import colors from '../../styles/colors'
-
-const PokemonDetail = (props) => {
-  return (
-    <Detail>
-      <Title>{props.title}</Title>
-      <Text>{props.detail}</Text>
-    </Detail>
-  );
-};
-
-
-PokemonDetail.propTypes = {
-  title: PropTypes.string,
-};
-
-export default PokemonDetail;
 
 // Styles
 
@@ -42,3 +23,9 @@ const Text = styled.p`
   font-size: 14px;
   width: 100%;
 `;
+
+export {
+  Detail,
+  Title,
+  Text
+}
