+++ src/components/pokemon/PokemonDetails.jsx
@@ -1,12 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styled from '@emotion/styled'
 
-import Detail from './PokemonDetail'
-import colors from '../../styles/colors'
-
+import Detail from './PokemonDetail.jsx'
+import { Wrapper, Details, Name, Title, Stats } from './PokemonDetails_styles'
 const PokemonDetails = ({data, showDetails}) => {
-  
   return (
     <Wrapper showDetails={showDetails}>
       <Details>
@@ -76,43 +73,3 @@ PokemonDetails.propTypes = {
 }
 
 export default PokemonDetails
-
-// styles
-
-const Wrapper = styled.div `
-  align-items: center;
-  display: ${props => props.showDetails === true ? 'flex' : 'none'};
-  flex-flow: row nowrap;
-  justify-content: space-between;
-  width: 100%;
-`;
-
-const Details = styled.div`
-  align-items: center;
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: space-evenly;
-  width: 50%;
-`;
-
-const Name = styled.h2`
-  color: ${colors.yellow};
-  font-size: 29px;
-  text-transform: uppercase;
-  margin: 15px 0px;
-`;
-
-const Stats = styled.div`
-  align-items: center;
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: space-evenly;
-  width: 50%;
-`;
-
-const Title = styled.h3`
-  color: ${colors.white};
-  font-size: 24px;
-  text-transform: uppercase;
-  margin: 15px 0px;
-`;
