+++ src/components/pokemon/PokemonImg_styles.js
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+// Styles
+
+const Picture = styled.picture``;
+
+const Img = styled.img``;
+
+export { 
+  Picture,
+  Img
+}
