+++ src/components/pokemon/Pokemon_styles.js
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+const Wrapper = styled.div`
+  align-items: center;
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: space-around;
+  width: 60%;
+`;
+
+export {
+  Wrapper
+}
