+++ src/components/searchBar/SearchBar_styles.js
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+const Form = styled.form`
+  align-items: center;
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: flex-start;
+  margin-bottom: 60px;
+`;
+
+export { Form }
